fix:refactor into components
