+++ src/App.js
@@ -1,13 +1,11 @@
 import { useState } from "react";
-
-const initialItems = [
-  { id: 1, description: "Passport", quantity: 2, packed: true },
-  { id: 2, description: "Socks", quantity: 10, packed: true },
-  { id: 3, description: "Charger", quantity: 1, packed: false },
-];
+import Logo from "./components/Logo";
+import Form from "./components/Form";
+import PackingList from "./components/PackingList";
+import Stats from "./components/Stats";
 
 function App() {
-  const [items, setItems] = useState(initialItems);
+  const [items, setItems] = useState([]);
 
   function handleAddItems(item) {
     setItems((items) => [...items, item]);
@@ -44,127 +42,3 @@ function App() {
   );
 }
 export default App;
-
-function Logo() {
-  return <h1>🌴 Far Away 💼</h1>;
-}
-
-function Form({ onAddItems }) {
-  const [description, setDescription] = useState("");
-  const [quantity, setQuantity] = useState(1);
-  const handleChange = (e) => {
-    const capitalized =
-      e.target.value.charAt(0).toUpperCase() +
-      e.target.value.slice(1).toLowerCase();
-    setDescription(capitalized);
-  };
-  function handleSubmit(e) {
-    e.preventDefault();
-    if (!description) return;
-    const newItem = { description, quantity, packed: false, id: Date.now() };
-    onAddItems(newItem);
-    setDescription("");
-    setQuantity(1);
-  }
-  return (
-    <form className="add-form" onSubmit={handleSubmit}>
-      <h3>What do you need for your 😍 trip?</h3>
-      <select
-        value={quantity}
-        onChange={(e) => setQuantity(Number(e.target.value))}
-      >
-        {Array.from({ length: 20 }, (_, i) => i + 1).map((num) => (
-          <option value={num} key={num}>
-            {num}
-          </option>
-        ))}
-      </select>
-      <input
-        type="text"
-        placeholder="Item..."
-        value={description}
-        onChange={handleChange}
-      />
-      <button>Add</button>
-    </form>
-  );
-}
-
-function PackingList({ items, onDeleteItems, onToogleItems, onClearList }) {
-  const [sortBy, setSortBy] = useState("input");
-  let sortedItems;
-
-  if (sortBy === "input") sortedItems = items;
-  if (sortBy === "description")
-    sortedItems = items
-      .slice()
-      .sort((a, b) => a.description.localeCompare(b.description));
-  if (sortBy === "packed")
-    sortedItems = items
-      .slice()
-      .sort((a, b) => Number(a.packed) - Number(b.packed));
-
-  return (
-    <div className="list">
-      <ul>
-        {sortedItems.map((item) => (
-          <Item
-            item={item}
-            onDeleteItems={onDeleteItems}
-            onToogleItems={onToogleItems}
-            key={item.id}
-          />
-        ))}
-      </ul>
-
-      <div className="actions">
-        <select value={sortBy} onChange={(e) => setSortBy(e.target.value)}>
-          <option value="input">Sort by input</option>
-
-          <option value="description">Sort by description</option>
-
-          <option value="packed">Sort by packed status</option>
-        </select>
-        <button onClick={onClearList}>Clear List</button>
-      </div>
-    </div>
-  );
-}
-
-function Item({ item, onDeleteItems, onToogleItems }) {
-  return (
-    <li>
-      <input
-        type="checkbox"
-        checked={item.packed}
-        onChange={() => onToogleItems(item.id)}
-      />
-      <span className={item.packed ? "packed" : ""}>
-        {item.quantity} {item.description}
-      </span>
-      <button onClick={() => onDeleteItems(item.id)}>❌</button>
-    </li>
-  );
-}
-
-function Stats({ items }) {
-  if (!items.length)
-    return (
-      <p className="stats">
-        <em>Start adding some items to your packing list</em>
-      </p>
-    );
-  const numItems = items.length;
-  const packedItems = items.filter((item) => item.packed).length;
-  const percentage = Math.round((packedItems / numItems) * 100);
-  return (
-    <footer className="stats">
-      <em>
-        {percentage === 100
-          ? "You got everything you need! Ready to go !!"
-          : `You have ${numItems} items on your list, and you already packed
-        ${packedItems} (${percentage}%)`}
-      </em>
-    </footer>
-  );
-}
