+++ src/components/Item.js
@@ -1,0 +1,15 @@
+export default function Item({ item, onDeleteItems, onToogleItems }) {
+  return (
+    <li>
+      <input
+        type="checkbox"
+        checked={item.packed}
+        onChange={() => onToogleItems(item.id)}
+      />
+      <span className={item.packed ? "packed" : ""}>
+        {item.quantity} {item.description}
+      </span>
+      <button onClick={() => onDeleteItems(item.id)}>❌</button>
+    </li>
+  );
+}
