+++ src/components/Stats.js
@@ -1,0 +1,21 @@
+export default function Stats({ items }) {
+  if (!items.length)
+    return (
+      <p className="stats">
+        <em>Start adding some items to your packing list</em>
+      </p>
+    );
+  const numItems = items.length;
+  const packedItems = items.filter((item) => item.packed).length;
+  const percentage = Math.round((packedItems / numItems) * 100);
+  return (
+    <footer className="stats">
+      <em>
+        {percentage === 100
+          ? "You got everything you need! Ready to go !!"
+          : `You have ${numItems} items on your list, and you already packed
+        ${packedItems} (${percentage}%)`}
+      </em>
+    </footer>
+  );
+}
